+++ src/OpenTelemetry.Instrumentation.Http/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 * Sanitize `http.url` attribute. ([#1961](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1961))
+* Added `RecordException` to HttpClientInstrumentationOptions which allows
+  Exception to be reported as ActivityEvent.
 
 ## 1.0.0-rc3